+++ CHANGELOG.md
@@ -33,6 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/789>)
 - Fix auto-documentation for the data_format plugins
   (<https://github.com/openvinotoolkit/datumaro/pull/793>)
+- Fix CVAT for images 1.1 data format import at project level
+  (<https://github.com/openvinotoolkit/datumaro/pull/795>)
 
 ### Security
 - Add security.md file for the SDL